backend/pyproject.toml
[project]
name = "xoxoedu-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "authlib>=1.4",
    "httpx>=0.28",
    "itsdangerous>=2.2",
    "celery[redis]>=5.4",
    "flower>=2.0",
    "redis>=5.2",
    "resend>=2.4",
    "python-slugify>=8.0",
    "python-multipart>=0.0.20",
    "starlette>=0.41",
    "boto3>=1.36",
    "stripe>=11.0",
    "weasyprint>=62.0",
    "litellm>=1.55",
    "openai>=1.50",
    "pgvector>=0.3",
    "jinja2>=3.1",
    "tenacity>=9.0",
    "sse-starlette>=2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "anyio>=4.7",
    "factory-boy>=3.3",
    "freezegun>=1.5",
    "respx>=0.21",
    "ruff>=0.8",
    "mypy>=1.13",
    "sqlalchemy[mypy]>=2.0",
    "types-python-jose>=3.3",
    "moto[s3]>=5.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["alembic/"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "TCH"]
ignore = ["B008", "TC003"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["sqlalchemy.ext.mypy.plugin"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=xml"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
